XML parser has now own module
